requirements.txt
pandas>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0 